requirements.txt
pandas>=1.5.0
requests>=2.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
streamlit>=1.28.0
plotly>=5.15.0
python-dotenv>=1.0.0

numpy>=1.24.0
scikit-learn>=1.3.0